Fixed quota reduction bug in WaveManager
The base strength and the quota strength scaling factor
determine the value with which the decision will be made,
whether a robot can be supported by the quota or not.
However, the scaled strength is only meant to be used to
prevent many of the same and strong robots to appear. It
does not represent the actual strength.

Before this fix, the scaled strength has been deducted
from the quota, when actually the base strength should
be subtracted.