Restructured cooldown action abstraction and wizard movement
Cooldown actions have been restructured, so that also actions
that shall not be triggered by the user (because it is not a skill)
can be run with cooldown. E.g. taking damage. Others, that are
triggered by the wizard are classified as mutually exclusive and
get triggered by a dedicated executor that has to know which skill
is active.