Fix previous direction updated when not moved
The previous direction stored in the WizardMovement script
has been overwritten, although the wizard did not move.

That has been caused by using yield return instead of yield break
for an early exit. The first way yielded a result and directly
jumped back into the coroutine. The effect of exiting when the
update flag is not set did not occur.