Added score based action unlock and static score access
Components such as the ScoreModel have to be implemented as an
object, but actually should be statically available.
This commit provides static access after to such components.

That is used for statically accessing the game score with minimal
overhead for checking if the score has exceeded the minimum score
needed to unlock a certain skill.